.claude/worktrees/unruffled-wilson-d8c2b0/requirements.txt
# Python dependencies
Django>=5.2
openpyxl
reportlab
# For API integration with React
django-cors-headers
djangorestframework
requests
gunicorn
uvicorn
dj-database-url
psycopg2-binary
whitenoise
# Testing
pytest
pytest-django
